move session from core to domain
