+++ src/domain/session/index.ts
@@ -1,1 +1,2 @@
 export * from "./session-manager";
+export * from "./user-session";
