+++ src/presentation/conversations/appointment-conversation.ts
@@ -2,9 +2,9 @@ import { format } from "date-fns";
 
 import { Weekday } from "@/common/constants";
 import { TypeConvesations } from "@/domain/interfaces";
-import { UserSession } from "@/core/user-sesssion";
+import { UserSession } from "@/domain/session/user-session";
 import { IConversation } from "@/domain/usecases";
-import { TypeSend } from "../interfaces";
+import { TypeSend } from "../session";
 import Messages from "../messages";
 
 export class AppointmentConversation implements IConversation {
