+++ src/presentation/conversations/inform-cpf-conversation.ts
@@ -1,8 +1,8 @@
 import { TypeConvesations } from "@/domain/interfaces";
-import { UserSession } from "@/core/user-sesssion";
+import { UserSession } from "@/domain/session/user-session";
 import { PatientService } from "@/domain/services";
 import { IConversation } from "@/domain/usecases";
-import { TypeSend } from "../interfaces";
+import { TypeSend } from "../session";
 import Messages from "../messages";
 
 export class InformCpfConversation implements IConversation {
