+++ src/presentation/conversations/inform-specialty-conversation.ts
@@ -1,9 +1,9 @@
 import { TypeConvesations } from "@/domain/interfaces";
 import { ClinicModel } from "@/domain/models";
-import { UserSession } from "@/core/user-sesssion";
+import { UserSession } from "@/domain/session/user-session";
 import { CalendarService } from "@/domain/services";
 import { IConversation } from "@/domain/usecases";
-import { TypeSend } from "../interfaces";
+import { TypeSend } from "../session";
 import Messages from "../messages";
 
 export class InformSpecialtyConversation implements IConversation {
