--- src/presentation/interfaces/index.ts
@@ -1,2 +1,0 @@
-export * from "@/presentation/interfaces/send-read";
-export * from "@/presentation/interfaces/context-manager";
