+++ src/presentation/session/context-manager.ts
@@ -1,4 +1,4 @@
-import { TypeRead, TypeSend } from "@/presentation/interfaces";
+import { TypeRead, TypeSend } from "@/presentation/session";
 
 export interface IContextManager {
   read: TypeRead;
